+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 0.14.16
+
+### Patch Changes
+
+-   Updated dependencies [fdfa554]
+    -   @sap-ux/ui5-application-writer@0.17.13
+
+## 0.14.15
+
+### Patch Changes
+
+-   Updated dependencies [964ab8e]
+    -   @sap-ux/ui5-application-writer@0.17.12
+
 ## 0.14.14
 
 ### Patch Changes